requirements.txt
absl-py==0.14.0
asgiref==3.4.1
ortools==9.0.9048
paho-mqtt==1.5.1
protobuf==3.18.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
